Refactor domain abilities into loader definitions
