+++ plugins/woocommerce/src/Internal/Abilities/Domain/OrderUpdateStatus.php
@@ -1,0 +1,111 @@
+<?php
+/**
+ * Order update status ability definition file.
+ */
+
+declare( strict_types=1 );
+
+namespace Automattic\WooCommerce\Internal\Abilities\Domain;
+
+use Automattic\WooCommerce\Internal\Abilities\AbilityDefinition;
+
+defined( 'ABSPATH' ) || exit;
+
+/**
+ * Registers the WooCommerce order update status ability.
+ */
+class OrderUpdateStatus extends DomainAbility implements AbilityDefinition {
+
+	/**
+	 * Register the ability.
+	 *
+	 * @since 10.9.0
+	 */
+	public static function register(): void {
+		if ( self::has_ability( 'woocommerce/order-update-status' ) ) {
+			return;
+		}
+
+		wp_register_ability(
+			'woocommerce/order-update-status',
+			array(
+				'label'               => __( 'Update Order Status', 'woocommerce' ),
+				'description'         => __(
+					'Update a WooCommerce order status using WooCommerce order APIs.',
+					'woocommerce'
+				),
+				'category'            => 'woocommerce',
+				'input_schema'        => self::get_input_schema(),
+				'output_schema'       => self::get_entity_output_schema( 'order' ),
+				'execute_callback'    => array( __CLASS__, 'execute' ),
+				'permission_callback' => array( __CLASS__, 'can_manage_orders' ),
+				'meta'                => self::get_domain_meta( 'update-status', false, false, true ),
+			)
+		);
+	}
+
+	/**
+	 * Update an order status.
+	 *
+	 * @param array $input Ability input.
+	 * @return array|\WP_Error
+	 *
+	 * @since 10.9.0
+	 */
+	public static function execute( array $input ) {
+		$order = self::get_order_from_input( $input );
+
+		if ( is_wp_error( $order ) ) {
+			return $order;
+		}
+
+		if ( empty( $input['status'] ) ) {
+			return new \WP_Error(
+				'woocommerce_order_status_required',
+				__( 'Order status is required.', 'woocommerce' ),
+				array( 'status' => 400 )
+			);
+		}
+
+		$order->update_status(
+			sanitize_key( $input['status'] ),
+			isset( $input['note'] ) ? self::sanitize_string( $input['note'] ) : ''
+		);
+
+		return array(
+			'order' => self::format_order( $order, false ),
+		);
+	}
+
+	/**
+	 * Check order management access.
+	 *
+	 * @param mixed $input Ability input.
+	 * @return bool
+	 *
+	 * @since 10.9.0
+	 */
+	public static function can_manage_orders( $input = array() ): bool {
+		$order_id = self::get_input_id( $input );
+
+		return $order_id > 0 && wc_rest_check_post_permissions( 'shop_order', 'edit', $order_id );
+	}
+
+	/**
+	 * Get the ability input schema.
+	 *
+	 * @return array
+	 */
+	private static function get_input_schema(): array {
+		return array(
+			'type'                 => 'object',
+			'properties'           => array(
+				'id'     => array( 'type' => 'integer' ),
+				'status' => array( 'type' => 'string' ),
+				'note'   => array( 'type' => 'string' ),
+			),
+			'required'             => array( 'id', 'status' ),
+			'additionalProperties' => false,
+		);
+	}
+}
